fix: book id types

+++ backend/src/main/java/com/bookstore/backend/repository/BooksRepository.java
@@ -3,5 +3,5 @@ package com.bookstore.backend.repository;
 import com.bookstore.backend.entity.Books;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BooksRepository extends JpaRepository<Books, Long> {
+public interface BooksRepository extends JpaRepository<Books, Integer> {
 }
